// src/modules/_shared/middlewares/error-code.enum.ts
const enum ErrorCode {
  // AUTH_EMAIL_ALREADY_EXISTS = "AUTH_EMAIL_ALREADY_EXISTS",
  // AUTH_INVALID_TOKEN = "AUTH_INVALID_TOKEN",
  // AUTH_USER_NOT_FOUND = "AUTH_USER_NOT_FOUND",
  // AUTH_NOT_FOUND = "AUTH_NOT_FOUND",
  // AUTH_TOO_MANY_ATTEMPTS = "AUTH_TOO_MANY_ATTEMPTS",
  // AUTH_UNAUTHORIZED_ACCESS = "AUTH_UNAUTHORIZED_ACCESS",
  // AUTH_TOKEN_NOT_FOUND = "AUTH_TOKEN_NOT_FOUND",

  // Access Control Errors
  FORBIDDEN = "FORBIDDEN",
  UNAUTHORIZED = "UNAUTHORIZED",

  // Validation and Resource Errors
  VALIDATION_ERROR = "VALIDATION_ERROR",
  NOT_FOUND = "NOT_FOUND",

  // System Errors
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  VERIFICATION_ERROR = "VERIFICATION_ERROR",

  // Additional Errors
  BAD_REQUEST = "BAD_REQUEST",
  CONFLICT = "CONFLICT",
  UNPROCESSABLE_ENTITY = "UNPROCESSABLE_ENTITY",
  SERVICE_UNAVAILABLE = "SERVICE_UNAVAILABLE",
}

export { ErrorCode };